refactor(nms): Separate the GatewayTierContext
